+++ src/libfm/src/Data.h
@@ -88,7 +88,7 @@ class Data {
 		LargeSparseMatrix<DATA_FLOAT>* data;
 		DVector<DATA_FLOAT> target;
 
-		uint64 num_feature;
+		uint num_feature;
 		uint num_cases;
  
 		DATA_FLOAT min_target;
@@ -188,7 +188,7 @@ void Data::load(std::string filename) {
 		}
 		DATA_FLOAT _value;
 		int nchar;
-        uint64 _feature;
+        uint _feature;
 		while (!fData.eof()) {
 			std::string line;
 			std::getline(fData, line);
@@ -200,7 +200,7 @@ void Data::load(std::string filename) {
 				min_target = std::min(_value, min_target);
 				max_target = std::max(_value, max_target);			
 				num_rows++;
-				while (sscanf(pline, "%llu:%f%n", &_feature, &_value, &nchar) >= 2) {
+				while (sscanf(pline, "%d:%f%n", &_feature, &_value, &nchar) >= 2) {
 					pline += nchar;	
 					num_feature = std::max(_feature, num_feature);
 					has_feature = true;
@@ -240,7 +240,7 @@ void Data::load(std::string filename) {
 		uint64 cache_id = 0;
 		DATA_FLOAT _value;
 		int nchar;
-        uint64 _feature;
+        uint _feature;
 		while (!fData.eof()) {
 			std::string line;
 			std::getline(fData, line);
@@ -254,7 +254,7 @@ void Data::load(std::string filename) {
 				data.value[row_id].data = &(cache[cache_id]);
 				data.value[row_id].size = 0;
 			
-				while (sscanf(pline, "%llu:%f%n", &_feature, &_value, &nchar) >= 2) {
+				while (sscanf(pline, "%d:%f%n", &_feature, &_value, &nchar) >= 2) {
 					pline += nchar;	
 					assert(cache_id < num_values);
 					cache[cache_id].id = _feature;
